+++ zetaclient/chains/evm/common/validation_test.go
@@ -1,4 +1,4 @@
-package evm_test
+package common
 
 import (
 	"fmt"
@@ -9,14 +9,12 @@ import (
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/onrik/ethrpc"
 	"github.com/stretchr/testify/require"
-
 	"github.com/zeta-chain/node/pkg/coin"
-	"github.com/zeta-chain/node/zetaclient/chains/evm"
 	"github.com/zeta-chain/node/zetaclient/testutils"
 )
 
 // the relative path to the testdata directory
-var TestDataDir = "../../"
+var TestDataDir = "../../../"
 
 func TestCheckEvmTxLog(t *testing.T) {
 	// test data
@@ -93,11 +91,11 @@ func TestCheckEvmTxLog(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			fmt.Printf("check test: %s\n", tt.name)
-			err := evm.ValidateEvmTxLog(
+			err := ValidateEvmTxLog(
 				tt.vLog,
 				connectorAddr,
 				"0xb252c9e77feafdeeae25cc1f037a16c4b50fa03c494754b99a7339d816c79626",
-				evm.TopicsZetaSent,
+				TopicsZetaSent,
 			)
 			if tt.fail {
 				require.Error(t, err)
@@ -293,7 +291,7 @@ func TestCheckEvmTransactionTable(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := evm.ValidateEvmTransaction(tt.tx)
+			err := ValidateEvmTransaction(tt.tx)
 			if tt.fail {
 				require.Error(t, err)
 				require.ErrorContains(t, err, tt.msg)
@@ -310,109 +308,109 @@ func TestCheckEvmTransaction(t *testing.T) {
 
 	t.Run("should pass for valid transaction", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.NoError(t, err)
 	})
 	t.Run("should fail for nil transaction", func(t *testing.T) {
-		err := evm.ValidateEvmTransaction(nil)
+		err := ValidateEvmTransaction(nil)
 		require.ErrorContains(t, err, "transaction is nil")
 	})
 	t.Run("should fail for empty hash", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.Hash = ""
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "hash is empty")
 	})
 	t.Run("should fail for negative nonce", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.Nonce = -1
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "nonce -1 is negative")
 	})
 	t.Run("should fail for empty from address", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.From = ""
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "not a valid hex address")
 	})
 	t.Run("should fail for invalid from address", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.From = "0x"
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "from 0x is not a valid hex address")
 	})
 	t.Run("should pass for empty to address", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.To = ""
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.NoError(t, err)
 	})
 	t.Run("should fail for invalid to address", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.To = "0xinvalid"
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "to 0xinvalid is not a valid hex address")
 	})
 	t.Run("should fail for negative value", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.Value = *big.NewInt(-1)
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "value -1 is negative")
 	})
 	t.Run("should fail for negative gas", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.Gas = -1
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "gas -1 is negative")
 	})
 	t.Run("should fail for negative gas price", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.GasPrice = *big.NewInt(-1)
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "gas price -1 is negative")
 	})
 	t.Run("should remove '0x' prefix from input data", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.NoError(t, err)
 		require.Equal(t, "", tx.Input)
 	})
 	t.Run("nil block number should pass", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.BlockNumber = nil
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.NoError(t, err)
 	})
 	t.Run("should fail for negative block number", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		negBlockNumber := -1
 		tx.BlockNumber = &negBlockNumber
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "block number -1 is not positive")
 	})
 	t.Run("should fail for empty block hash", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.BlockHash = ""
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "block hash is empty")
 	})
 	t.Run("nil transaction index should fail", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.TransactionIndex = nil
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "index is nil")
 	})
 	t.Run("should fail for negative transaction index", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		negTransactionIndex := -1
 		tx.TransactionIndex = &negTransactionIndex
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "index -1 is negative")
 	})
 	t.Run("should fail for invalid input data", func(t *testing.T) {
 		tx := testutils.LoadEVMInbound(t, TestDataDir, 1, inboundHash, coin.CoinType_Gas)
 		tx.Input = "03befinvalid"
-		err := evm.ValidateEvmTransaction(tx)
+		err := ValidateEvmTransaction(tx)
 		require.ErrorContains(t, err, "input data is not hex encoded")
 	})
 }