+++ changelog.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
-## unreleased
+## Unreleased
+
+### Features
+
+* [3353](https://github.com/zeta-chain/node/pull/3353) - add liquidity cap parameter to ZRC20 creation
 
 ### Features
 
@@ -12,6 +16,7 @@
 * [3360](https://github.com/zeta-chain/node/pull/3360) - update protocol contract imports using consolidated path
 * [3349](https://github.com/zeta-chain/node/pull/3349) - implement new bitcoin rpc in zetaclient with improved performance and observability
 * [3379](https://github.com/zeta-chain/node/pull/3379) - add Avalanche, Arbitrum and World Chain in chain info
+* [3390](https://github.com/zeta-chain/node/pull/3390) - orchestrator V2: EVM observer-signer
 
 ### Fixes